src/main/java/com/rarchives/ripme/App.java
package com.rarchives.ripme;

import java.util.ArrayList;
import java.util.List;
import org.apache.log4j.Logger;

public class App {
    public static Logger logger;
    public List<String> options = new ArrayList<>();
    public List<String> args = new ArrayList<>();

    public void handleArguments(String[] arguments) {
        for (String arg : arguments) {
            if (arg.startsWith("--")) {
                addOption(arg);
            } else {
                addArgument(arg);
            }
        }
    }

    private void addOption(String option) {
        options.add(option);
    }

    private void addArgument(String arg) {
        args.add(arg);
    }

    public void complexOperation(OperationParameters params) {
        // Complex operation using params
    }

    public static class OperationParameters {
        private int a, b, c, d, e, f, g;

        // Constructor, getters, setters
        public OperationParameters(int a, int b, int c, int d, int e, int f, int g) {
            this.a = a;
            this.b = b;
            this.c = c;
            this.d = d;
            this.e = e;
            this.f = f;
            this.g = g;
        }

        public int getA() { return a; }
        public int getB() { return b; }
        public int getC() { return c; }
        public int getD() { return d; }
        public int getE() { return e; }
        public int getF() { return f; }
        public int getG() { return g; }
    }

    public void performOperation() {
        // Method implementation
    }

    public void performAllOperations() {
        new Operation1().execute();
        new Operation2().execute();
        new Operation3().execute();
    }

    private class Operation1 {
        void execute() {
            // Operation 1 implementation
        }
    }

    private class Operation2 {
        void execute() {
            // Operation 2 implementation
        }
    }

    private class Operation3 {
        void execute() {
            // Operation 3 implementation
        }
    }

    public static void main(String[] args) {
        App app = new App();
        app.handleArguments(args);
    }
}
